engine: Extract SimulationProgress into simulation_progress.hpp
